// src/app/models/transaction.ts
export class Transaction{
    
    time: Date;
    user: String;
    chassisNo:String;
    km:Number;
    transaction: String;
    comment: String;
    // constructor(

    // time: Date,
    // user: String,
    // chassisNo:String,
    // km:Number,
    // transaction: String,
    // comment: String
    // ){}
}